motor-fw: Change controller's speed 'inc' field to 16-bit.
Allowing more creative uses.

+++ boards/motor/firmware/trunk/control.c
@@ -23,7 +23,7 @@ typedef struct
 		bool enabled;
 
 		/* How much to increment the position by */
-		uint8_t inc;
+		uint16_t inc;
 		/* How often to increment the position */
 		uint16_t period;
